Fix wrong nuget package dependency
If you generate two projects in this solution that publish nuget
packages, and one depends on the other, without the
GetPackageVersionDependsOn change the second project will always depend
on version 1.0.0 of the first.

This took a lot of searching and trial and error to find out.

See https://stackoverflow.com/questions/52165586/nuget-task-ignore-custom-version-of-referenced-projects

The no-build flag was to allow packing the dependency project (was
failing before).